Add validation-rules support and extend public API
Introduce full validation-rules support, implemented in the new
erbsland::conf::vr namespace and built as optional CMake targets
(variant without regex, with std::regex, or with Erbsland RegEx).

Extend documentation (reference + tutorials) and modernize
documentation build setup. Add static library installation workflow.

Public API extensions:
- New types: CaseSensitivity, Matrix<T>, StringList, ValueList,
  ConstValueList, ValueMatrix, ConstValueMatrix, StringConvertible,
  StringLike
- Extended Value, Parser, Name, NamePath, String, Error, Bytes,
  ValueType, and ValueIterator APIs
- ValueList now std::vector<ValuePtr>; add ConstValueList
- RegEx constructor now takes multiline flag; equality includes it
- Deprecate ValueType::isSingle() (use isScalar())
- Minor: Signer now takes SignatureSignerPtr by value

+++ src/erbsland/conf/TimeOffset.cpp
@@ -41,7 +41,7 @@ auto TimeOffset::seconds() const noexcept -> int {
     if (isLocalTime()) {
         return 0;
     }
-    return std::abs(_seconds.count()) % 60;
+    return static_cast<int>(std::abs(_seconds.count()) % 60);
 }
 
 
@@ -49,7 +49,7 @@ auto TimeOffset::minutes() const noexcept -> int {
     if (isLocalTime()) {
         return 0;
     }
-    return (std::abs(_seconds.count()) / 60) % 60;
+    return static_cast<int>((std::abs(_seconds.count()) / 60) % 60);
 }
 
 
@@ -57,7 +57,7 @@ auto TimeOffset::hours() const noexcept -> int {
     if (isLocalTime()) {
         return 0;
     }
-    return std::abs(_seconds.count()) / 3600;
+    return static_cast<int>(std::abs(_seconds.count()) / 3600);
 }
